Add files via upload
Office Space is a WebGL project that utilizes 3d models to simulate an office scene. 

To run this project. Close out of all chrome tabs. Open the CMD prompt in Administrator mode. Locate the URL of the chrome.exe application.
CD to the chrome application folder. Type the following command: chrome.exe --allow-file-access-from-files. Locate the Project folder and open it. The previous steps should have enabled the file textures in Chrome.